requirements.txt
flask>=2.3.0
redis>=4.5.0
mecab-python3>=1.0.6
jaconv>=0.3.4
openai>=1.0.0
pydantic>=2.0.0
gunicorn>=21.0.0
